requirements.txt

Django
djangorestframework
django-cors-headers
djangorestframework
psycopg2-binary
pillow
gunicorn
djongo
django-cloudinary-storage